[FIX] mrp: update test_order.py
- Add docstrings
- Use Form views in all tests
- Add some missing assets
